[scudo] Reduce the scope of AllocAfterFork
`ScudoWrappersCppTest.AllocAfterFork` was failing obscurely sometimes.
Someone pointed us to Linux's `vm.max_map_count` that can be
significantly lower on some machines than others. It turned out that
on a machine with that setting set to 65530, some `ENOMEM` errors
would occur with `mmap` & `mprotect` during that specific test.

Reducing the number of times we fork, and the maximum size allocated
during that test makes it pass on those machines.

Differential Revision: https://reviews.llvm.org/D111342

GitOrigin-RevId: 6727832c324c1fb43946275d24e2931fde94bc0d
Change-Id: I02ee6f28c90366dd63171690df4ef8f82fc62289
Signed-off-by: spezi77 <spezi7713@gmx.net>